+++ framework/db/Schema.php
@@ -15,6 +15,7 @@ use yii\base\NotSupportedException;
 use yii\caching\Cache;
 use yii\caching\CacheInterface;
 use yii\caching\TagDependency;
+use yii\helpers\StringHelper;
 
 /**
  * Schema is the base class for concrete DBMS-specific schema classes.
@@ -90,6 +91,19 @@ abstract class Schema extends BaseObject
     public $columnSchemaClass = 'yii\db\ColumnSchema';
 
     /**
+     * @var string|string[] character used to quote schema, table, etc. names.
+     * An array of 2 characters can be used in case starting and ending characters are different.
+     * @since 2.0.14
+     */
+    protected $tableQuoteCharacter = "'";
+    /**
+     * @var string|string[] character used to quote column names.
+     * An array of 2 characters can be used in case starting and ending characters are different.
+     * @since 2.0.14
+     */
+    protected $columnQuoteCharacter = '"';
+
+    /**
      * @var array list of ALL schema names in the database, except system schemas
      */
     private $_schemaNames;
@@ -105,6 +119,10 @@ abstract class Schema extends BaseObject
      * @var QueryBuilder the query builder for this database
      */
     private $_builder;
+    /**
+     * @var string server version as a string.
+     */
+    private $_serverVersion;
 
 
     /**
@@ -509,7 +527,12 @@ abstract class Schema extends BaseObject
      */
     public function quoteSimpleTableName($name)
     {
-        return strpos($name, "'") !== false ? $name : "'" . $name . "'";
+        if (is_string($this->tableQuoteCharacter)) {
+            $startingCharacter = $endingCharacter = $this->tableQuoteCharacter;
+        } else {
+            list($startingCharacter, $endingCharacter) = $this->tableQuoteCharacter;
+        }
+        return strpos($name, $startingCharacter) !== false ? $name : $startingCharacter . $name . $endingCharacter;
     }
 
     /**
@@ -521,7 +544,48 @@ abstract class Schema extends BaseObject
      */
     public function quoteSimpleColumnName($name)
     {
-        return strpos($name, '"') !== false || $name === '*' ? $name : '"' . $name . '"';
+        if (is_string($this->tableQuoteCharacter)) {
+            $startingCharacter = $endingCharacter = $this->columnQuoteCharacter;
+        } else {
+            list($startingCharacter, $endingCharacter) = $this->columnQuoteCharacter;
+        }
+        return $name === '*' || strpos($name, $startingCharacter) !== false ? $name : $startingCharacter . $name . $endingCharacter;
+    }
+
+    /**
+     * Unquotes a simple table name.
+     * A simple table name should contain the table name only without any schema prefix.
+     * If the table name is not quoted, this method will do nothing.
+     * @param string $name table name.
+     * @return string unquoted table name.
+     * @since 2.0.14
+     */
+    public function unquoteSimpleTableName($name)
+    {
+        if (is_string($this->tableQuoteCharacter)) {
+            $startingCharacter = $this->tableQuoteCharacter;
+        } else {
+            $startingCharacter = $this->tableQuoteCharacter[0];
+        }
+        return strpos($name, $startingCharacter) === false ? $name : substr($name, 1, -1);
+    }
+
+    /**
+     * Unquotes a simple column name.
+     * A simple column name should contain the column name only without any prefix.
+     * If the column name is not quoted or is the asterisk character '*', this method will do nothing.
+     * @param string $name column name.
+     * @return string unquoted column name.
+     * @since 2.0.14
+     */
+    public function unquoteSimpleColumnName($name)
+    {
+        if (is_string($this->columnQuoteCharacter)) {
+            $startingCharacter = $this->columnQuoteCharacter;
+        } else {
+            $startingCharacter = $this->columnQuoteCharacter[0];
+        }
+        return strpos($name, $startingCharacter) === false ? $name : substr($name, 1, -1);
     }
 
     /**
@@ -606,6 +670,19 @@ abstract class Schema extends BaseObject
     {
         $pattern = '/^\s*(SELECT|SHOW|DESCRIBE)\b/i';
         return preg_match($pattern, $sql) > 0;
+    }
+
+    /**
+     * Returns a server version as a string comparable by [[\version_compare()]].
+     * @return string server version as a string.
+     * @since 2.0.14
+     */
+    public function getServerVersion()
+    {
+        if ($this->_serverVersion === null) {
+            $this->_serverVersion = $this->db->getSlavePdo()->getAttribute(\PDO::ATTR_SERVER_VERSION);
+        }
+        return $this->_serverVersion;
     }
 
     /**