+++ framework/db/mysql/Schema.php
@@ -11,6 +11,7 @@ use yii\base\InvalidConfigException;
 use yii\base\NotSupportedException;
 use yii\db\ColumnSchema;
 use yii\db\Constraint;
+use yii\db\ConstraintFinderInterface;
 use yii\db\ConstraintFinderTrait;
 use yii\db\Exception;
 use yii\db\Expression;
@@ -25,7 +26,7 @@ use yii\helpers\ArrayHelper;
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0
  */
-class Schema extends \yii\db\Schema
+class Schema extends \yii\db\Schema implements ConstraintFinderInterface
 {
     use ConstraintFinderTrait;
 
@@ -69,6 +70,15 @@ class Schema extends \yii\db\Schema
         'varbinary' => self::TYPE_BINARY,
         'json' => self::TYPE_JSON,
     ];
+
+    /**
+     * @inheritDoc
+     */
+    protected $tableQuoteCharacter = '`';
+    /**
+     * @inheritDoc
+     */
+    protected $columnQuoteCharacter = '`';
 
     /**
      * @inheritDoc
@@ -193,28 +203,6 @@ SQL;
     protected function loadTableDefaultValues($tableName)
     {
         throw new NotSupportedException('MySQL does not support default value constraints.');
-    }
-
-    /**
-     * Quotes a table name for use in a query.
-     * A simple table name has no schema prefix.
-     * @param string $name table name
-     * @return string the properly quoted table name
-     */
-    public function quoteSimpleTableName($name)
-    {
-        return strpos($name, '`') !== false ? $name : "`$name`";
-    }
-
-    /**
-     * Quotes a column name for use in a query.
-     * A simple column name has no prefix.
-     * @param string $name column name
-     * @return string the properly quoted column name
-     */
-    public function quoteSimpleColumnName($name)
-    {
-        return strpos($name, '`') !== false || $name === '*' ? $name : "`$name`";
     }
 
     /**