+++ tests/data/postgres.sql
@@ -27,6 +27,7 @@ DROP TABLE IF EXISTS "T_constraints_4";
 DROP TABLE IF EXISTS "T_constraints_3";
 DROP TABLE IF EXISTS "T_constraints_2";
 DROP TABLE IF EXISTS "T_constraints_1";
+DROP TABLE IF EXISTS "T_upsert";
 
 DROP SCHEMA IF EXISTS "schema1" CASCADE;
 DROP SCHEMA IF EXISTS "schema2" CASCADE;
@@ -332,4 +333,17 @@ CREATE TABLE "T_constraints_4"
     "C_col_1" INT NULL,
     "C_col_2" INT NOT NULL,
     CONSTRAINT "CN_constraints_4" UNIQUE ("C_col_1", "C_col_2")
+);
+
+CREATE TABLE "T_upsert"
+(
+    "id" SERIAL NOT NULL PRIMARY KEY,
+    "ts" INT NULL,
+    "email" VARCHAR(128) NOT NULL UNIQUE,
+    "recovery_email" VARCHAR(128) NULL,
+    "address" TEXT NULL,
+    "status" SMALLINT NOT NULL DEFAULT 0,
+    "orders" INT NOT NULL DEFAULT 0,
+    "profile_id" INT NULL,
+    UNIQUE ("email", "recovery_email")
 );